[TASK] Mitigate extbase crash in non TS context
People tend to run extbase in context it is not
designed to be run in. TYPO3 v11 and before
has been forgiving in such cases and falls back
to a broken (empty) TypoScript configution.

TYPO3 v12 is more picky in this regard, and throws
an exception. With current TYPO3 v12 frontend
bootstrap, extension developers are a bit limited
in their options to solve this.

The patch adds a hack to extbase bootstrap code
to be as forgiving as TYPO3 v11 has been. However,
the case is logged as deprecated, and a comment
is added to explain the scenario and options.

Resolves: #99417
Related: #97816
Releases: main
Change-Id: I0bbe3bdfe1dd942e2aecd633cc19f907b739bdac
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78007
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>