[TASK] Streamline ObjectMonitoring- and DomainObjectInterface usage
ObjectMonitoringInterface is a signaling interface for extbase
internals to signal that an object state can be monitored. Its
main purpose is to tell the persistence manager if an object or
an object collection is so called dirty, i.e. if changes need to
be persisted or not.

This change lets DomainObjectInterface extend ObjectMonitoringInterface
instead of AbstractDomainObject. Abstract classes must never be external
api, they should just contain boilerplate/repetative code. With this
change extbase can perform actions based on implemented interfaces, i.e.
guaranteed functionality instead of checking object instances.

Releases: main
Resolves: #100069
Change-Id: I65548f3fe3f6a51c2b87e67c35642c431e2d7bb2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78012
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>