[TASK] Simplify ValidatorResolver::buildBaseValidatorConjunction()
Prior to this change, buildBaseValidatorConjunction() was responsible
for initializing $this->baseValidatorConjunctions[$indexKey] and
additionally adding an object validator if the given target type
was a class name.

Now, the initialization takes place in getBaseValidatorConjunction()
already and buildBaseValidatorConjunction() is only called if target
types is a class name.

That way, buildBaseValidatorConjunction() gets rid of an if-statement
and the whole method looses one indentation.

Additionally, buildBaseValidatorConjunction() is guaranteed to get
passed a class-string, so argument name $targetClassName actually
makes sense.

Releases: main
Resolves: #100191
Change-Id: I7003fa1e827274e0a1eed3674e675a6f30a6c07c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78142
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>