[TASK] Deprecate TCA config MM_insert_fields
"True" MM relations (intermediata table with uid_local
and uid_foreign columns, and no TCA for this table) are
still hard to configure in TYPO3. One reason is the
amount of TCA config options: It's often not clear
to developers which ones need to be set, and which ones
are optional or useless.

One of the 'useless' options is MM_insert_fields: It
allows setting a column in the intermediate table to a
configured value. Intermediate tables are transparent
in TYPO3: They are fully managed by the system, there
is not way to edit them within the framework. As such,
having an additional column within intermediate tables
that carry some value, is useless.

Looking at usages within extensions of MM_insert_fields,
it turns out developers typically confuse it with
MM_match_fields (this config is a story on its own).
Developers simply tend to set both MM_insert_fields and
MM_match_fields to the same value "to be sure".

The patch deprecates MM_insert_fields without
substitution in v12. TYPO3 v13 will actively remove
this config option from TCA.

Resolves: #100335
Releases: main
Change-Id: I7832321fb3bcbfc00bb6f288c6f0cd5ceb41dc64
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78297
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>