[TASK] Mark all Qom classes and interfaces internal
Some of these classes have been marked @internal
already. Extending / changing this area of the
extbase framework should not bee needed, and is
rather hard to achieve as well in order to have
some effect.

The patch adds @internal to all classes in this
namespace that don't have this annotation, yet.

Releases: main
Resolves: #100375
Change-Id: Id083a1fa9c4d4eaff0fc3d321aca5eb17192f098
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78502
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>