[TASK] Deprecate ConfigurationManager->getContentObject()
The extbase ConfigurationManager logic is still
painful. It is a stateful singleton, which
we can not dissolve at the moment.

To reduce headaches, we now deprecate the
getContentObject() misuse, a solution avoiding
this has been established with #100623.

This allows us to streamline the implementation
with TYPO3 v13 quite a bit, making it less of a
burden.

The patch deprecates the method and introduces a
setRequest() method instead. Extbase bootstrap
takes care of setting this for each plugin. We
don't get rid of the singleton state, but since
we don't add a public getRequest(), it makes
ConfigurationManager a data sink only for this
state, which can not be abused anymore.

This also allows us to get rid of
$GLOBALS['TYPO3_REQUEST'] usages within
ConfigurationManager in TYPO3 v13.

Resolves: #100662
Related: #100623
Releases: main
Change-Id: I419b5008d80b6545c5fe55067af6031845978680
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78731
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>