[TASK] Use psr/http-message:"^1.1 || ^2.0"
We should strive for current dependency versions
with upcoming 12.4.0 LTS.

psr/http-message has recently been released
as 1.1.0 and 2.0.0, which we should support.

According to [1], having ^1.1 as lower bound is
encouraged to have ^2.0.

All our implementations are @internal, dropping
1.0 shouldn't be a huge burden, and other
dependencies like guzzle adapted as well already.

The patch bumps a couple of psr dependencies to
allow ^2.0, raises guzzlehttp/psr7 to a minimum
version that supports ^1.1 and ^2.0, and sets
psr/http-message to ^1.1 || ^2.0.

Our implementations are adapted, allowing us to
be much more strict on types.

One unfortunate "nullable" Uri is left, but this
could be sorted out with a dedicated patch since
it has impact on quite a few test cases.

[1] https://www.php-fig.org/psr/psr-7/meta/#72-type-additions

> composer u psr/http-client \
             psr/http-factory \
             psr/http-server-handler \
             psr/http-server-middleware
> composer req guzzlehttp/psr7:^2.5.0
> composer req --no-update guzzlehttp/psr7:"^2.5.0" -d typo3/sysext/core
> composer req psr/http-message:"^1.1 || ^2.0"
> composer req --no-update psr/http-message:"^1.1 || ^2.0" -d typo3/sysext/adminpanel
> composer req --no-update psr/http-message:"^1.1 || ^2.0" -d typo3/sysext/core
> composer req --no-update psr/http-message:"^1.1 || ^2.0" -d typo3/sysext/form
> composer req --no-update psr/http-message:"^1.1 || ^2.0" -d typo3/sysext/redirects

Resolves: #100682
Releases: main
Change-Id: I98bb4031a681fe65827e282e0af059af66863873
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78763
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>