[TASK] Allow doctrine/instantiator ^1.5 || ^2.0
v11 has "^1.4". For v12, we first removed it entirely,
and brought doctrine/instantiator back as "^2.0"
in #100207.

The "^2.0" restriction is a bit harsh, not all
other libraries caught up on this relatively
young new major.

To play well with others, we change
doctrine/instantiator to "^1.5 || ^2.0", the changes
betwen 1.5 and 2.0 [1] are mainly type
sugar that is not critical to us. The ^1.5 minor
restriction shouldn't be an issue for "sane" other
libraries which tend to go with something like "^1.0",
or "^1.3", or similar.

> composer req doctrine/instantiator:"^1.5 || ^2.0"
> composer req --no-update doctrine/instantiator:"^1.5 || ^2.0" -d typo3/sysext/extbase

[1] https://github.com/doctrine/instantiator/compare/1.5.0...2.0.0

Resolves: #100630
Related: #100207
Releases: main
Change-Id: I9b4de7b57a89a6db65e4fb9154196b3790c50104
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78779
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>