[TASK] Remove usages of QueryBuilder::resetQueryParts
Doctrine 4 will not allow any manipulation to query parts, what makes
it necessary to remove and replace the places where query parts
are manipulated throughout the core. As a rule of thumb there
should be no reusage of queryBuilder instances anyway, as this
can easily lead to problems with conflicting configurations.

In case of resetQueryParts the transition is rather easy, because only
one test file uses it, and it can easily be replaced with
dedicated instances of the queryBuilder.

Resolves: #100900
Releases: main
Change-Id: I4bdadd177e6d52a6af6e10dc65c218ece5dab22e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79092
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>