[TASK] Clean up usages of ContentObjectRenderer request attribute
ContentObjectRenderer is now added as request attribute in
certain cases. But since the signature of obtaining an attribute
from a request correctly states, that the attribute can be null,
consuming code should account for that and defensively
fall back to defaults, when the object can not be obtained.

Releases: 12.4, main
Resolves: #101201
Change-Id: I3b3145794cafd8f5a8c1081e97599f9862be8b99
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80921
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>