[TASK] Update naming in MvcPropertyMappingConfigurationService
With #88682 Extbase argument validation has switched from using PHP
serialize/unserialize to use json_encode/json_decode. The doc
comments and variables in the responsible class
`MvcPropertyMappingConfigurationService` however still use the
term "serialize", which is misleading and not correct anymore.

This change removes any occurrence of the term "serialize" and
replaces it with a term that suits better to the current
situation.

Resolves: #101563
Related: #88682
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I0306880c636e5f6e6153829407c2451a91c23907
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80324
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>