[TASK] Prevent Extbase errorAction from writing session data
Any validation error and the resulting FlashMessage items that is
handled implicitly by the Extbase ActionController::errorAction
will be persisted to a user session. In case the session does not
exist, it will be generated and a new session cookie is sent.

This change aims to make "persisting to session storage" explicit
and to transfer errorAction FlashMessages internally.

Resolves: #102906
Releases: main
Change-Id: Ife3efc6b527583fa0e067e836fc6be051d64e448
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82605
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Oliver Hader <oliver.hader@typo3.org>