[BUGFIX] Use locale dependencies in Extbase LocalizationUtility
Since TYPO3 v12.4, TYPO3 uses locales such as "de_DE"
or "de-AT" in Site Configuration natively.

In order to resolve the "locale dependencies" natively,
a new Locale object needs to be created be the "Locales"
factory class. This is needed to check for "de_AT.locallang.xlf"
and falling back to "de.locallang.xlf" in a multi-step mode.

This change adapts one place in Extbase's LocalizationUtility
where a new Locale object was previously instantiated via
new Locale() instead of using the "Locales" class which
deals with all dependencies (incl. manually defined dependencies)
directly there.

A new comment is added to explain this inside the constructor
of Locale.

In addition, some more tests are added (including the ones
that did not work before), as well as an example that
it does not matter if the file is called "de-AT.locallang.xlf"
(with a hyphen) or "de_CH.locallang.xlf" (with an underscore)

Resolves: #102579
Related: #101710
Releases: main, 12.4
Change-Id: Id51082e75cf635bbe3cbdfc906dc2a7442ab06af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82065
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>