[BUGFIX] Use matching site in extbase BE modules
Extbase BE modules suffer from the conceptual mistake
of using FE TypoScript for configuration. Modules
not in page context pick the first root template
in the page tree as TS entry point.

The patch changes this awful handling to also use
the attached site object for this page to substitute
site settings used as constants in TS correctly.

Resolves: #102452
Related: #97816
Releases: main, 12.4
Change-Id: I24e1e814cd20fc4999311d4f0068fb51818ad611
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82558
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>