[TASK] Drop the `AbstractFileFolder` superclass
This internal superclass provided methods with types that were
overwritten with a different signature in subclasses, which means
that the abstraction wass not helpful (and prevented using proper
native type declarations). The superclass is hence now dropped.

As the dropped class had been marked as `@internal`, this change
is not breaking.

The `AbstractFileFolderConverter` needs some more cleanup, but
that is out of scope for this change.

Resolves: #102884
Related: #102879
Releases: main
Change-Id: I11fe0388f28685c16bb86b3beb8606fe90e1834e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82532
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>