[TASK] Remove internal declaration for hashService property in actionController
With #92996 the property `hashService` in extbase `actionController`
has been marked as internal. This change was most likely made,
because the extbase `hashService` class was declared as internal.

The extbase `HashService` has been deprecated with #102763 and
extbase `actionController` now uses the new `HashService` introduced
with #102761, which is public API. Therefore, the internal declaration
for the property `hashService` in `actionController` can be removed
again, so extension authors can use the `hashService` for e.g. HMAC
validation.

Resolves: #103320
Related: #92996
Related: #102763
Related: #102761
Releases: main
Change-Id: I6ded62172953bd61e676f367eec04bd20526feed
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83303
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>