[BUGFIX] Fix return type of Query::execute()
`\TYPO3\CMS\Extbase\Persistence\Generic\Query::execute()` has
a boolean flag to define if the result should be a raw result
without data mapping to an object. In this case a raw array
is returned and not an array with objects. The annotation
has been fixed at several places to be in sync with underlying
\Doctrine\DBAL\Result::fetchAllAssociative() return type.

Resolves: #103874
Releases: main, 12.4
Change-Id: Iea3eda9e805ca064d290f088d6130a44c06c1a42
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84343
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>