[BUGFIX] Adapt ObjectConverter test with new symfony/property-info
With symfony property-info 7.1.2, the type resolving for
collections has been hardened.

https://github.com/symfony/property-info/commit/d0bbc495f11ab99a1e2cbf04699331b88b3c1293

Our extbase ObjectConverter test for collections uses an anonymous
class with an ObjectStorage property. The annotation for the property
as well as the native property type is however not a FQCN. So the
anonymous class has no information about, what FQCN the
`ObjectStorage` class is.

The change adds the FQCN for the ObjectStorage in the
anonymous class, which resolves the failing test, and
adds a test with a non-anymous class to test short
namespaces still work.

Resolves: #104254
Releases: main, 12.4, 11.5
Change-Id: Idc64e2533e0bc6d2dccc9724643247772f4f5d11
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85000
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>