[BUGFIX] Avoid ancient injectSettings() handling
There is some ancient when a DI method is called
'injectSettings': DI ignores this.

There is a single use for Fluid views in extbase,
which calls injectSettings if it exists manually.
However no Fluid view ever implemented this method.
It is removed.

Also, the DI magic to not call 'injectSettings'
is removed, injectSettings can now be used as
casual inject* method. This is considered a
bugfix (for main only): It is unexpected and
undocumented this handling exists.

Resolves: #104445
Releases: main
Change-Id: I6eb527bea048dd856f618b4caadb7e17d4ce6ad3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85317
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>