[TASK] Avoid action magic in getControllerActionName()
Extbase has some code within ExtbaseRequestParameters
->getControllerActionName() that fetches all methods
from the given controller class to see if the
previously set action name is in there. The comparison
includes some case insensitivity magic and in the
end tries to return an action name which is cased
exactly as the method within the controller.

This code has been added to flow3 back in 2010 and
was back ported to extbase around that time. It was
however never relevant in extbase: The extbase
RequestBuilder has an earlier step where it
compares the requested action with the list of
allowed registered actions (via ConfigurationManager)
and throws an exception if there is no
(case sensitive!) match.

The simplifies getControllerActionName() by
removing this obsolete code.

Resolves: #104472
Related: #104471
Releases: main
Change-Id: I3219fb257923c298ac38393c53949af82e11ed19
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85347
Tested-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>