[TASK] Tame extbase ConfigurationManager
The extbase ConfigurationManager is - even after various
refactoring loops - still an anti pattern: This is a
stateful singleton initialized by extbase bootstrap with
state, and it can be injected into arbitrary classes.

This patch still does not solve the root of this issue:
It is still unclear on how to solve it.

The patch however now makes the two "worker" classes
stateless: Both the "concrete" implementation classes
FrontendConfigurationManager and BackendConfigurationManager.
This strategy bundles the ConfigurationManager flaws into
one class now. The patch also moves an access to
$GLOBALS['TYPO3_REQUEST'] out of __construct(), which
allows us to deprecate this call with an upcoming patch.

Note there is also TYPO3\CMS\Extbase\Service\ExtensionService,
which is *another* stateful injectable singleton, that works
on ConfigurationManager - #messDoesNotEndHere

Resolves: #104514
Releases: main
Change-Id: I5ef864b116fdbe285a64bdf60175fa0ad6fc780a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85384
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>