[BUGFIX] Provide uploaded files as extbase arguments
This patch enables property mapping of PSR-7
UploadedFile objects to DTO properties.

The workaround for this bug in EXT:form is reverted,
because it is not needed anymore.

Also, a previously broken argument naming for uploaded
files triggers a dedicated exception now.

Finally, the UploadedFile implementation of TYPO3 is
no longer marked as @internal, as it is now exposed to
extbase property mapping.

Resolves: #102283
Releases: main
Change-Id: I356073ab01505d99205a5dd960ca7dced2d5fb67
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85449
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Helmut Hummel <typo3@helhum.io>