[TASK] Use fluid attribute for Request handling
ext:fluid RenderingContext is in the bad habit of extending
fluid standalone RenderingContext with the two methods
setRequest() and getRequest() that are not part of fluid
standalone RenderingContextInterface.

This is ugly with type hinting and can't be changed in a
good way due to liskov collisions on method signatures in
view helpers.

Fluid standalone also does not want to add these methods
itself since it - as a view - wants to avoid dependencies
to PSR-7.

Fluid standalone instead added setAttribute(), hasAttribute()
and getAttribute() to RenderingContextInterface. This allows
adding "arbitrary" objects.

The patch makes use of this: On the setter side, the
RenderingContextFactory->create() is extended to accept
the request and set it as attribute, on the has/getter side
all consumers are adapted.

This deprecates setRequest() and getRequest(), which are
changed to transparently work with the attribute instead.

Resolves: #104684
Releases: main
Change-Id: Id096073765c00e06e0437539f963f59a69e3fba0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85708
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>