[TASK] Deprecate ext:fluid fillDefaultsByPackageName()
TemplatePaths->fillDefaultsByPackageName() is located
at the wrong place by having this magic within ext:fluid:
It is the last piece of ext:fluid core (non-VH) classes
that mixes extbase logic into fluid.

The method is only triggered actively by extbase, and
indirectly via parent __construct() in standalone fluid
if the class is constructed with a string, which is not
used by the core.

The patch refactors the single consumer in extbase
ActionController, which leads to simplified fluid path
setup in that class. Two obsolete heavy mocking tests
that don't add value bite the dust along the way.

Even though ext:fluid TemplatePaths is marked @internal,
the method is not removed directly but deprecated since
it is used by quite a few extensions in practice, and
it does not harm much to just let it log deprecation
note in v13.

Resolves: #104764
Releases: main
Change-Id: Ia2fbcb38a6ef76f6aad41fcda927e3f32447ca9d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85813
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Simon Praetorius <simon@praetorius.me>