[TASK] Improve extbase SourceInterface handling
We stumbled upon the fishy SourceInterface when working
on #104990: This is an empty interface, extended by
SelectorInterface and JoinInterface (violating interface
composition over inheritance), and implemented by
Join and Selector respectively.

The patch changes SelectorInterface and JoinInterface to
no longer extend SourceInterface, and implements both
in Join / Selector.

Looking at consumers, various types are added to add
more native PHP type support to the extbase QOM layer.
This substitutes some type commments, which turned out
to be partially wrong: Most notably selector() of
QueryObjectModelFactory, which is called with
$nodeTypeName being null.

This codebase should receive a bigger overhaul, and
it helps to have more types we can rely on when doing so.

Resolves: #105049
Related: #104990
Releases: main
Change-Id: I95a5a1382cfd5e6dff24de7ee733a4749e435150
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86201
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>