[BUGFIX] Properly set columnMap relation for type "folder"
TCA type "folder" is a special type since it's always
a relation but not to a table but one or more folder
identifier. This is now handled properly.

In previous versions, this was handled (by accident?)
since the $type (the FQN) matched a strpbrk() check,
which led to marking the relation as "toOne".

Resolves: #105317
Releases: main, 13.4
Change-Id: Ib5e1e5f1762801b43c619bd1fe9a95b3339f8ffa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86682
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>