[BUGFIX] Respect model type configuration in ColumnMapFactory
With #104808, the `ColumnMapFactory` was refactored
to rely on the `TcaSchema` and the associated
`RelationshipType` for building the ColumnMap.
Especially setting a `Relation`. While it's
generally correct to rely on TCA, the type
definition in a model had previously influence
on the resolving in the `ColumnMapFactory `.

An example is TCA type `passthrough`, which
is no `RelationalFieldTypeInterface`, as it
does not define any relation. However it's
possible and common in user land to assign
such property a dedicated type or a
`ObjectStroage` in the model, which
makes the field an actual relation.

For those use cases the extbase model
should continue to overrule the TCA
definition. Firstly for backwards
compatibility and secondly to reduce
the need for extbase extension developers
to fiddle around with TCA settings.

Resolves: #105062
Related: #104808
Releases: main, 13.4
Change-Id: I7da1c146ad8dc4cc2f17da38a77c977106357fcc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86936
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>