[BUGFIX] Fix missing CType evaluation for UriBuilder
The method "ExtensionService->getTargetPidByPlugin()" only
evaluated the combination of tt_content.CType = 'list' and
tt_content.list_type = 'pluginSignature' when preparing
the retrieval of plugin.tx_myext.view.defaultPid.

With #105538 this was fixed for TYPO3 v14 already,
and allows to properly evaluate CType scenarios.

However, for TYPO3 v13 (where "list_type" is still
allowed and only deprecated) this CType scenario was
not evaluated.

Tests are added for this behavior, and a fix is
performed to allow retrieving both ways
in the QueryBuilder based logic.

Resolves: #105865
Releated: #105538
Releases: 13.4
Change-Id: Id394ef11bb8d8989a15c2503bb094a9688262558
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87597
Reviewed-by: Sebastian Iffland <sebastian.iffland@fullhaus.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>